Remove unused state
