+++ .storybook/manager-head.html
@@ -1,0 +1,1 @@
+<meta name="viewport" content="width=device-width, initial-scale=1">
